+++ tests/g2p.cpp
@@ -43,9 +43,9 @@ TEST(g2p_a, linear_bspline) {
     ASSERT_EQ(myMPM.g_ngridz(), 16) << "ngridz not calculated correctly";
 
     // setup grid
-    for (size_t i = 0 ; i < myMPM.g_ngridx(); ++i) 
-        for (size_t j = 0; j < myMPM.g_ngridy(); ++j)
-            for (size_t k = 0; k < myMPM.g_ngridz(); ++k) {
+    for (int i = 0 ; i < myMPM.g_ngridx(); ++i) 
+        for (int j = 0; j < myMPM.g_ngridy(); ++j)
+            for (int k = 0; k < myMPM.g_ngridz(); ++k) {
                 myMPM.g_momentumx(i, j, k) = 0.1*(i+j+k);
                 myMPM.g_momentumy(i, j, k) = 0.2*(i+j+k);
                 myMPM.g_momentumz(i, j, k) = 0.3*(i+j+k);
@@ -64,7 +64,7 @@ TEST(g2p_a, linear_bspline) {
     // check conservation
     // sum particles' force (m*a) and momentum (dxdt*m)
     double psum[6] {0., 0., 0., 0., 0., 0.}, gsum[6] {0., 0., 0., 0., 0., 0.};
-    for (size_t i = 0; i < myMPM.p_size(); ++i) {
+    for (int i = 0; i < myMPM.p_size(); ++i) {
         psum[0] += myMPM.p_ax(i)*myMPM.p_mass(i);
         psum[1] += myMPM.p_ay(i)*myMPM.p_mass(i);
         psum[2] += myMPM.p_az(i)*myMPM.p_mass(i);
@@ -72,7 +72,7 @@ TEST(g2p_a, linear_bspline) {
         psum[4] += myMPM.p_dydt(i)*myMPM.p_mass(i);
         psum[5] += myMPM.p_dzdt(i)*myMPM.p_mass(i);
     }
-    for (size_t i = 0; i < myMPM.g_size(); ++i) {
+    for (int i = 0; i < myMPM.g_size(); ++i) {
         gsum[0] += myMPM.g_forcex(i);
         gsum[1] += myMPM.g_forcey(i);
         gsum[2] += myMPM.g_forcez(i);
@@ -100,9 +100,9 @@ TEST(g2p_a, cubic_bspline) {
     ASSERT_EQ(myMPM.g_ngridz(), 18) << "ngridz not calculated correctly";
 
     // setup grid
-    for (size_t i = 0 ; i < myMPM.g_ngridx(); ++i) 
-        for (size_t j = 0; j < myMPM.g_ngridy(); ++j)
-            for (size_t k = 0; k < myMPM.g_ngridz(); ++k) {
+    for (int i = 0 ; i < myMPM.g_ngridx(); ++i) 
+        for (int j = 0; j < myMPM.g_ngridy(); ++j)
+            for (int k = 0; k < myMPM.g_ngridz(); ++k) {
                 myMPM.g_momentumx(i, j, k) = 0.1*(i+j+k);
                 myMPM.g_momentumy(i, j, k) = 0.2*(i+j+k);
                 myMPM.g_momentumz(i, j, k) = 0.3*(i+j+k);
@@ -121,7 +121,7 @@ TEST(g2p_a, cubic_bspline) {
     // check conservation
     // sum particles' force (m*a) and momentum (dxdt*m)
     double psum[6] {0., 0., 0., 0., 0., 0.}, gsum[6] {0., 0., 0., 0., 0., 0.};
-    for (size_t i = 0; i < myMPM.p_size(); ++i) {
+    for (int i = 0; i < myMPM.p_size(); ++i) {
         psum[0] += myMPM.p_ax(i)*myMPM.p_mass(i);
         psum[1] += myMPM.p_ay(i)*myMPM.p_mass(i);
         psum[2] += myMPM.p_az(i)*myMPM.p_mass(i);
@@ -129,7 +129,7 @@ TEST(g2p_a, cubic_bspline) {
         psum[4] += myMPM.p_dydt(i)*myMPM.p_mass(i);
         psum[5] += myMPM.p_dzdt(i)*myMPM.p_mass(i);
     }
-    for (size_t i = 0; i < myMPM.g_size(); ++i) {
+    for (int i = 0; i < myMPM.g_size(); ++i) {
         gsum[0] += myMPM.g_forcex(i);
         gsum[1] += myMPM.g_forcey(i);
         gsum[2] += myMPM.g_forcez(i);
@@ -157,9 +157,9 @@ TEST(g2p_strainspinrates, linear_bspline) {
     ASSERT_EQ(myMPM.g_ngridz(), 16) << "ngridz not calculated correctly";
 
     // setup grid
-    for (size_t i = 0; i < myMPM.g_ngridx(); ++i)
-        for (size_t j = 0; j < myMPM.g_ngridy(); ++j) 
-            for (size_t k = 0; k < myMPM.g_ngridz(); ++k) {
+    for (int i = 0; i < myMPM.g_ngridx(); ++i)
+        for (int j = 0; j < myMPM.g_ngridy(); ++j) 
+            for (int k = 0; k < myMPM.g_ngridz(); ++k) {
                 const double x = i*0.2 + myMPM.g_mingridx();
                 const double y = j*0.2 + myMPM.g_mingridy();
                 const double z = k*0.2 + myMPM.g_mingridz();
@@ -210,9 +210,9 @@ TEST(g2p_strainspinrates, cubic_bspline) {
     ASSERT_EQ(myMPM.g_ngridz(), 18) << "ngridz not calculated correctly";
 
     // setup grid
-    for (size_t i = 0; i < myMPM.g_ngridx(); ++i)
-        for (size_t j = 0; j < myMPM.g_ngridy(); ++j) 
-            for (size_t k = 0; k < myMPM.g_ngridz(); ++k) {
+    for (int i = 0; i < myMPM.g_ngridx(); ++i)
+        for (int j = 0; j < myMPM.g_ngridy(); ++j) 
+            for (int k = 0; k < myMPM.g_ngridz(); ++k) {
                 const double x = i*0.2 + myMPM.g_mingridx();
                 const double y = j*0.2 + myMPM.g_mingridy();
                 const double z = k*0.2 + myMPM.g_mingridz();
